install dialog renaming
